% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_exported_utils.R
\name{bb_log}
\alias{bb_log}
\title{log to the local bigbrithr database}
\usage{
bb_log(logger, label)
}
\arguments{
\item{logger}{logging function}

\item{label}{label}
}
\description{
log to the local bigbrithr database
}
