% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_exported_loggers.R
\name{bb_message}
\alias{bb_message}
\title{bb_message}
\usage{
bb_message(..., domain = NULL, appendLF = TRUE, label = "")
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or (for
    \code{message} only) a single condition object.}

\item{domain}{see \code{\link[base]{gettext}}.  If \code{NA}, messages will
    not be translated, see also the note in \code{\link[base]{stop}}.}

\item{appendLF}{logical: should messages given as a character string
    have a newline appended?}

\item{label}{label in bigbrothr database}
}
\description{
bbfied message
}
