% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_exported_loggers.R
\name{bb_stop}
\alias{bb_stop}
\title{bb_stop}
\usage{
bb_stop(..., call. = TRUE, domain = NULL, label = "")
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{call.}{logical, indicating if the call should become part of the
    error message.}

\item{domain}{see \code{\link[base]{gettext}}.  If \code{NA}, messages will
    not be translated.}

\item{label}{label in bigbrothr database}
}
\description{
bbfied stop
}
