% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_exported_utils.R
\name{bbfy}
\alias{bbfy}
\title{change a function into a bigbrothr loging function}
\usage{
bbfy(f, nm, label_arg = "label")
}
\arguments{
\item{f}{the function to bbfy}

\item{nm}{the name of the function to create}

\item{label_arg}{the name of the label}
}
\description{
As a side effect, a function named \code{nm} is created in the local environment.
}
